requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
transformers
numpy
pydantic
matplotlib
seaborn
tqdm